Jira issue HF-13
This was a fix for a bug whereby PDFs were not being uploaded. The fix was to alter settings in php.ini, which is not under version control.
I also removed some unnecessary debug code form file.class.php, but i am tagging this with HF-13 in order to be consistent