Jira issue HF-14
This is an issue whereby newly created assets are not created along with an array of variant sizes, as expected.
This fix involved using sizeify or php-imageMagick, as appropriate.
Apache rewriting is being used to create needed assets lazily. It is triggered by a custom 404 script, written in to vhost.conf